requirements.txt
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.3
blinker==1.6.2
click==8.1.7
Flask==2.3.3
frozendict==2.3.8
graphql-core==3.2.3
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
libcst==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mypy-extensions==1.0.0
Pygments==2.16.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.0
rich==13.5.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.0.post1
strawberry-graphql==0.205.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.5.0
uvicorn==0.23.2
Werkzeug==2.3.7
zipp==3.16.2
